requirements.txt
numpy==2.1.1
pillow==10.4.0
scikit-learn==1.5.2
django==5.1.1
djangorestframework==3.15.2